set dns options in bootstrap
