+++ config/envoyconfig/clusters_test.go
@@ -539,14 +539,7 @@ func Test_buildCluster(t *testing.T) {
 					"typedConfig": {
 						"@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
 						"dnsLookupFamily": "V4_ONLY",
-						"respectDnsTtl": true,
-						"typedDnsResolverConfig": {
-							"name": "envoy.network.dns_resolver.cares",
-							"typedConfig": {
-								"@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-								"udpMaxQueries": 100
-							}
-						}
+						"respectDnsTtl": true
 					}
 				},
 				"connectTimeout": "10s",
@@ -614,14 +607,7 @@ func Test_buildCluster(t *testing.T) {
 					"typedConfig": {
 						"@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
 						"dnsLookupFamily": "V4_PREFERRED",
-						"respectDnsTtl": true,
-						"typedDnsResolverConfig": {
-							"name": "envoy.network.dns_resolver.cares",
-							"typedConfig": {
-								"@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-								"udpMaxQueries": 100
-							}
-						}
+						"respectDnsTtl": true
 					}
 				},
 				"connectTimeout": "10s",
@@ -978,14 +964,7 @@ func Test_buildCluster(t *testing.T) {
 					"typedConfig": {
 						"@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
 						"dnsLookupFamily": "V4_ONLY",
-						"respectDnsTtl": true,
-						"typedDnsResolverConfig": {
-							"name": "envoy.network.dns_resolver.cares",
-							"typedConfig": {
-								"@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-								"udpMaxQueries": 100
-							}
-						}
+						"respectDnsTtl": true
 					}
 				},
 				"connectTimeout": "10s",
