+++ config/envoyconfig/testdata/clusters.json
@@ -38,14 +38,7 @@
       "typedConfig": {
         "@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
         "dnsLookupFamily": "V4_PREFERRED",
-        "respectDnsTtl": true,
-        "typedDnsResolverConfig": {
-          "name": "envoy.network.dns_resolver.cares",
-          "typedConfig": {
-            "@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-            "udpMaxQueries": 100
-          }
-        }
+        "respectDnsTtl": true
       }
     },
     "loadAssignment": {
@@ -98,14 +91,7 @@
       "typedConfig": {
         "@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
         "dnsLookupFamily": "V4_PREFERRED",
-        "respectDnsTtl": true,
-        "typedDnsResolverConfig": {
-          "name": "envoy.network.dns_resolver.cares",
-          "typedConfig": {
-            "@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-            "udpMaxQueries": 100
-          }
-        }
+        "respectDnsTtl": true
       }
     },
     "circuitBreakers": {
@@ -172,14 +158,7 @@
       "typedConfig": {
         "@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
         "dnsLookupFamily": "V4_PREFERRED",
-        "respectDnsTtl": true,
-        "typedDnsResolverConfig": {
-          "name": "envoy.network.dns_resolver.cares",
-          "typedConfig": {
-            "@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-            "udpMaxQueries": 100
-          }
-        }
+        "respectDnsTtl": true
       }
     },
     "circuitBreakers": {
@@ -246,14 +225,7 @@
       "typedConfig": {
         "@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
         "dnsLookupFamily": "V4_PREFERRED",
-        "respectDnsTtl": true,
-        "typedDnsResolverConfig": {
-          "name": "envoy.network.dns_resolver.cares",
-          "typedConfig": {
-            "@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-            "udpMaxQueries": 100
-          }
-        }
+        "respectDnsTtl": true
       }
     },
     "circuitBreakers": {
@@ -316,14 +288,7 @@
       "typedConfig": {
         "@type": "type.googleapis.com/envoy.extensions.clusters.dns.v3.DnsCluster",
         "dnsLookupFamily": "V4_PREFERRED",
-        "respectDnsTtl": true,
-        "typedDnsResolverConfig": {
-          "name": "envoy.network.dns_resolver.cares",
-          "typedConfig": {
-            "@type": "type.googleapis.com/envoy.extensions.network.dns_resolver.cares.v3.CaresDnsResolverConfig",
-            "udpMaxQueries": 100
-          }
-        }
+        "respectDnsTtl": true
       }
     },
     "circuitBreakers": {
